v2.0.0
 - fully automated TLS certificate management, just put your
   domain into /etc/hostname, and the system takes care of the rest
 - fully automated swap file generator runit/autorun/mkswap,
   improves system stability in critical low memory conditions
 - add: runit/bootup/11-printinfo.sh for extra system info during boot
 - fix: sources.list Devuan repository link
 - fix: hardened autorun functionality (disabled symlink support)
 - fix: pureftpd runscript now correctly handles VPC/LAN/WAN environment
 - other small fixes

+++ etc/runit/bootup/10-cleanup.sh
@@ -1,15 +1,11 @@
 # *-*- Shell Script -*-*
 # from VOID Linux (https://www.voidlinux.eu)
 
-msg "Cleanup..."
+msg "Cleanup ..."
 
 install -m0664 -o root -g utmp /dev/null /run/utmp
-if [ ! -e /var/log/wtmp ]; then
-	install -m0664 -o root -g utmp /dev/null /var/log/wtmp
-fi
-if [ ! -e /var/log/btmp ]; then
-	install -m0600 -o root -g utmp /dev/null /var/log/btmp
-fi
+[ -e /var/log/wtmp ] || install -m0664 -o root -g utmp /dev/null /var/log/wtmp
+[ -e /var/log/btmp ] || install -m0600 -o root -g utmp /dev/null /var/log/btmp
 #install -dm1777 /tmp/.X11-unix /tmp/.ICE-unix
 rm -f /etc/nologin /forcefsck /forcequotacheck /fastboot
